# design/ClientApplication/RegisterAgent/RegisterAgentOpenApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Register Agent",
    "description": "This API is an intermerdiary between an author agent and the Hyperledger Fabric blockchain application. It receives calls from the author agent and sends the petitions to the corresponding chaincodes.",
    "version": "0.0.1"
  },
  "externalDocs": {
    "description": "Github Repository",
    "url": "https://github.com/sfl0r3nz05/DistributedBchFWArchIoT"
  },
  "servers": [
    {
      "url": "To be hosted"
    }
  ],
  "tags": [
    {
      "name": "register",
      "description": "Register the firmware update or an author."
    }
  ],
  "paths": {
    "/register/author": {
      "post": {
        "tags": [
          "register"
        ],
        "summary": "Register the author",
        "description": "Receives a RegisterPetition from the author agent, and calls the registerAuthor chaincode from the blockchain application.",
        "operationId": "registerAuthor",
        "requestBody": {
          "description": "It contains a RegisterPetition",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "PublicKey": {
                    "type": "string"
                  },
                  "SignedMessage": {
                    "type": "string"
                  },
                  "Message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation. Returns the registerKey generated by the chaincode.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "The blockchain application found an error. Error message contained.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "register"
        ],
        "summary": "Register firmware update",
        "description": "Author agent sends an UpdateRegister object. The registerUpdate chaincode is called",
        "operationId": "register",
        "requestBody": {
          "description": "Contains an UpdateRegister object.",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "AuthorKey": {
                    "type": "string"
                  },
                  "Manifest": {
                    "type": "object"
                  },
                  "Payload": {
                    "type": "string",
                    "format": "binary"
                  },
                  "AuthorSign": {
                    "type": "string"
                  },
                  "AuthorManifestSign": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation."
          },
          "403": {
            "description": "The blockchain application found an error. Error message contained.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  }
}